Add HTTPClient leveraging Combine and Laziness
Due to the nature of the app as a CLI tool, it is necessary for the API calls to be blocking.

For composability and general ease of reasoning it is also useful to be able to treat them reactively, using Combine.

This also allows us to leverage Combine’s Deferred<>, allowing the HTTPClient to defer the actual API call until the Future is materialised/subscribed to.