Add ReadDeviceInfoCommand
This implements the very simple API call to get information for a specific device registered to the team. It only has simple output.

For the moment, its output is done inside the command struct; but when the renderer refactor lands, we will move this and the list devices output functions.