Remove `print` from completion renderer
Generally speaking, completed tasks should only output their results, and nothing else. Unless there’s been an error, we don’t want to change the output in any way. This means we can take the rendered result from STDOUT, be it JSON, YML or what have you, and use it right away. We should expect the output of our program to become the input to another, as yet unknown, program.