Add ResultRenderer
Previously, every command had to have a `output` function. These functions basically all do the same thing: depending on the output format, encode the results, then print it.

Now that we have `Renderer`, we can make a new one for rendering the output for all our commands. The simple one, which I have made here, is just what it was doing before. It encodes the result as JSON, YAML, or as a table and prints it.

Each model type will need to know how to render itself in a table, however. This is done with `ResultRenderable` conformance on the models.