Implement ListAppsCommand with filters
Previously, `ListAppsCommand` was only stubbed out. It would only dump out the response from the API request. It also used `DispatchGroup` to await the results.

Now, it’s formatted like the other commands; and we are using the deferred futures on the `HTTPClient`.

It also has some basic filtering options.