Introduce CSV output format with `CodableCSV`
Add a new output format: CSV. We have brought in the `CodableCSV` package to help us.

Despite its name, `CodableCSV`, does not have an `Encoder` for making CSVs. Instead, I have re-used the `TableInfoProvider` protocol to provide its `CSVWriter` type with the data it needs. This gets us CSV support with relatively little effort.

In future, if `CodableCSV` is updated to support `Encodable`, we could possibly revisit this and make encoding a CSV more like encoding JSON or YAML.