Create ReadBundleIdCommand
This is a multi-step command. It does two requests.

The read bundle ID information takes an opaque bundle ID identifier. This identifier is not displayed in the `list` command; it is internal to App Store Connect. We fetch it first, in the `findOpaqueIdentifier(for:)` func. When that returns, we can use that to make the request to read the bundle id.

https://developer.apple.com/documentation/appstoreconnectapi/read_bundle_id_information