Make sync remove and ignore invitations
Previously, we were only looking at the users list, and inviting people who were not on it. This causes trouble when run twice. If we had invited someone, but they have not yet accepted their invitation, the code would attempt to invite them again. That is not allowed by the API. It would return an error and the program would end.

Now, we look at not only who is in the users list, but also who has been invited. This has made the 'diffing' code a lot more complex, but it allows us to compare across two lists.

For everyone in the input list, we check if they are in either the existing users and invited users lists. If they are there, we can ignore that entry. If they are not, we can invite them.

For everyone in the existing users list that we want to remove, there are two things we might have to do: remove them from the users list, or remove them from the invited users list. Since the users and invited users are separate API calls there are two separate operations for deleting users or uninviting them.